Add worker review stats

+++ updates/create_worker_reviews_table.php
@@ -17,14 +17,14 @@ class CreateWorkerReviewsTable extends Migration
 
             $table->boolean('is_visible')->default(false);
             $table->integer('user_id')->unsigned()->index()->nullable();
-            $table->integer('rating')->nullable();
+            $table->decimal('rating', 3, 2)->nullable();
             $table->text('comment')->nullable();
             $table->text('breakdown')->nullable();
             $table->boolean('is_recommend')->default(true);
 
             $table->boolean('client_is_visible')->default(false);
             $table->integer('client_user_id')->unsigned()->index()->nullable();
-            $table->integer('client_rating')->nullable();
+            $table->decimal('client_rating', 3, 2)->nullable();
             $table->text('client_comment')->nullable();
 
             $table->string('invite_name')->nullable();
